starting to change functions to general forms. \n changed prototypes only. + changed int -> uint64_t where required.

+++ src/combinatorics.c
@@ -1,11 +1,11 @@
 #include "combinatorics.h"
 
-int choose(int n, int k)
+uint64_t choose(uint64_t n, uint64_t k)
 {
     return flooredFact(n, k) / fact(k);
 }
 
-int flooredFact(int n, int k)
+uint64_t flooredFact(uint64_t n, uint64_t k)
 {
     if (k == 0) return 1;
     int p = 1;
@@ -16,7 +16,7 @@ int flooredFact(int n, int k)
     return n * p;
 }
 
-int fact(int n)
+uint64_t fact(uint64_t n)
 {
     if (n == 0) return 1;
     int p = 1;
